pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "cws2"
description = "Tool for conlangers to document their languages"
authors = [{name = "m5ka", email = "git@marzka.cafe"}]
dynamic = ["version"]
requires-python = ">= 3.10"
dependencies = [
    "Django >= 4.1, < 4.2",
    "django-environ >= 0.10, < 0.11",
    "django-extensions >= 3.2.1, < 3.3",
    "django-gravatar2 >= 1.4.4, < 1.5",
    "django-jinja >= 2.10.2, < 2.11",
    "django-recaptcha >= 3.0, < 3.1",
    "django-sass >= 1.1, < 1.2",
    "fontawesomefree >= 6.3, < 6.4",
    "gunicorn >= 20.1, < 20.2",
    "mysqlclient >= 2.1.1, < 2.2",
    "python-slugify >= 8.0, < 8.1"
]

[project.optional-dependencies]
dev = [
    "black >= 23.1, < 23.2",
    "django-debug-toolbar >= 3.8.1, < 3.9",
    "faker >= 17.6, < 17.7",
    "flake8 >= 6.0, < 6.1",
]
test = [
    "pytest >= 7.2.2, < 7.3",
    "pytest-cov >= 4.0, < 4.1",
    "pytest-django >= 4.5.2, < 4.6"
]

[project.urls]
Home = "https://github.com/m5ka/cws2"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "cws2.settings"